BackEnd #17 : Created a seeder to post all the products at once
Created a seeder using which we can push all the products in products.json at once. If there is any anomaly in the database regarding multiple products, then we can just fix them in the products.json file and run the seeder, which will first delete every product details in the database and save again.