BackEnd #11 : Created a error handler class and setup a middleware for it
Previously we had set a fairly simple message for handling error, but now we are setting up this error handler class so that we can differentiate between development and production level error messages. In the development mode, we need to know the specifically where the error is occurring and why it is occurring, whereas in the production level, we are just interested to show the client a general message describing the error in short.