BackEnd #11 : Handled validation errors
Previously if two or more details about a product was missing while creating/updating a new product, then error of only one product was being thrown back, but now it is fixed. Also fixed a minor bug due to which all the product details were not being able to be fetched from the database.